Little clean up
