+++ clickerSolution.sln
@@ -1,0 +1,37 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.31005.135
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "clicker", "clicker\clicker.csproj", "{BF651993-C761-44C2-889F-F5D6EDC222CC}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TestProject", "TestProject\TestProject.csproj", "{A79ED5A9-DC88-4DF8-8A4B-68642E9625F5}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GameSystemObjects", "GameSystemObjects\GameSystemObjects.csproj", "{D2EF7614-036E-42DE-BDD1-CE0426C80D7A}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{BF651993-C761-44C2-889F-F5D6EDC222CC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{BF651993-C761-44C2-889F-F5D6EDC222CC}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{BF651993-C761-44C2-889F-F5D6EDC222CC}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{BF651993-C761-44C2-889F-F5D6EDC222CC}.Release|Any CPU.Build.0 = Release|Any CPU
+		{A79ED5A9-DC88-4DF8-8A4B-68642E9625F5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A79ED5A9-DC88-4DF8-8A4B-68642E9625F5}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A79ED5A9-DC88-4DF8-8A4B-68642E9625F5}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A79ED5A9-DC88-4DF8-8A4B-68642E9625F5}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D2EF7614-036E-42DE-BDD1-CE0426C80D7A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D2EF7614-036E-42DE-BDD1-CE0426C80D7A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D2EF7614-036E-42DE-BDD1-CE0426C80D7A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D2EF7614-036E-42DE-BDD1-CE0426C80D7A}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {011400BB-432B-469D-BDB4-F84A4CBF6C89}
+	EndGlobalSection
+EndGlobal
